drm: move file around so the distinction between main files and utils make more sense
