+++ src/core/decrypt/utils/__tests__/clean_old_stored_persistent_info.test.ts
@@ -48,8 +48,8 @@ const emptyPersistentSessionsStore = {
 function checkNothingHappen(persistentSessionsStore : any, limit : number) {
   const deleteLastSpy = jest.spyOn(persistentSessionsStore, "deleteOldSessions");
   const logInfoSpy = jest.fn();
-  jest.mock("../../../log", () => ({ __esModule: true as const,
-                                     default: { info: logInfoSpy } }));
+  jest.mock("../../../../log", () => ({ __esModule: true as const,
+                                        default: { info: logInfoSpy } }));
   const cleanOldStoredPersistentInfo = require("../clean_old_stored_persistent_info")
     .default;
   cleanOldStoredPersistentInfo(persistentSessionsStore, limit);
@@ -73,8 +73,8 @@ function checkRemoved(
 ) {
   const deleteLastSpy = jest.spyOn(persistentSessionsStore, "deleteOldSessions");
   const logInfoSpy = jest.fn();
-  jest.mock("../../../log", () => ({ __esModule: true as const,
-                                     default: { info: logInfoSpy } }));
+  jest.mock("../../../../log", () => ({ __esModule: true as const,
+                                        default: { info: logInfoSpy } }));
   const cleanOldStoredPersistentInfo = require("../clean_old_stored_persistent_info")
     .default;
   cleanOldStoredPersistentInfo(persistentSessionsStore, limit);
