+++ src/core/decrypt/utils/get_drm_system_id.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import startsWith from "../../utils/starts_with";
+import startsWith from "../../../utils/starts_with";
 
 /**
  * @param {string} keySystem
